# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Soccer Game OpenApi Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "vasudevan3092@gmial.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/v1/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get list of teams",
                "description": "Returns list of teams",
                "operationId": "getTeamsList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "path",
                        "description": "Search Term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "No Of items per page, if not mentioned all the data is dumped",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Store new Team",
                "description": "Returns team data",
                "operationId": "storeTeam",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTeamRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Teams"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/v1/teamplayers": {
            "get": {
                "tags": [
                    "Team Players"
                ],
                "summary": "Get list of teams",
                "description": "Returns list of team players",
                "operationId": "getTeamsPlayersList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "path",
                        "description": "Search Term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "team",
                        "in": "path",
                        "description": "Team vise players list",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "No Of items per page, if not mentioned all the data is dumped",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamPlayerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Team Players"
                ],
                "summary": "Store new Team Player",
                "description": "Returns player data",
                "operationId": "storeTeam",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTeamPlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamPlayers"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/teamplayers/{id}": {
            "get": {
                "tags": [
                    "Team Players"
                ],
                "summary": "Get player information",
                "description": "Returns Team player data",
                "operationId": "getTeamPlayersById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamPlayers"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Team Players"
                ],
                "summary": "Update existing Team Player",
                "description": "Returns updated team player data",
                "operationId": "updateTeamPlayer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTeamPlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamPlayers"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Team Players"
                ],
                "summary": "Delete existing Team Player",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTeamplayers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TeamPlayers": {
                "title": "team",
                "description": "team model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "firstName": {
                        "title": "firstName",
                        "description": "First Name of the new player",
                        "type": "string",
                        "example": "James"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "Last Name of the new player",
                        "type": "string",
                        "example": "Doe"
                    },
                    "playerImage": {
                        "title": "playerImage",
                        "description": "Upload a Player Image",
                        "type": "string",
                        "example": "Upload image for the team player"
                    },
                    "team_id": {
                        "title": "team_id",
                        "description": "Team's id of the new player",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "team"
                }
            },
            "Teams": {
                "title": "team",
                "description": "team model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new team",
                        "type": "string",
                        "example": "A nice team"
                    },
                    "logo": {
                        "title": "Logo",
                        "description": "Upload a Logo",
                        "type": "string",
                        "format": "binary",
                        "example": "Upload image logo for the team"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "team"
                }
            },
            "TeamPlayerResource": {
                "title": "TeamPlayerResource",
                "description": "Team Players resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeamPlayers"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TeamPlayerResource"
                }
            },
            "TeamResource": {
                "title": "TeamResource",
                "description": "Team resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Teams"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TeamResource"
                }
            },
            "StoreTeamPlayerRequest": {
                "title": "Store Teams request",
                "description": "Store Teams request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "firstName": {
                        "title": "firstName",
                        "description": "First Name of the new player",
                        "type": "string",
                        "example": "James"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "Last Name of the new player",
                        "type": "string",
                        "example": "Doe"
                    },
                    "playerImage": {
                        "title": "playerImage",
                        "description": "Upload a Player Image",
                        "type": "string",
                        "example": "Upload image for the team player"
                    },
                    "team_id": {
                        "title": "team_id",
                        "description": "Team's id of the new player",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "StoreTeamRequest": {
                "title": "Store Teams request",
                "description": "Store Teams request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new project",
                        "type": "string",
                        "example": "A nice project"
                    },
                    "logo": {
                        "title": "logo",
                        "description": "Upload a Logo (JPEG,JPG,PNG)",
                        "type": "string",
                        "example": "Upload image logo for the team "
                    }
                },
                "type": "object"
            },
            "UpdateTeamPlayerRequest": {
                "title": "Update Teams request",
                "description": "Update Teams request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "firstName": {
                        "title": "firstName",
                        "description": "First Name of the new player",
                        "type": "string",
                        "example": "James"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "Last Name of the new player",
                        "type": "string",
                        "example": "Doe"
                    },
                    "playerImage": {
                        "title": "playerImage",
                        "description": "Upload a Player Image",
                        "type": "string",
                        "example": "Upload image for the team player"
                    },
                    "team_id": {
                        "title": "team_id",
                        "description": "Team's id of the new player",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        []
    ]
}